# backend-laravel/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApiPost - linuxitos",
        "description": "API con login, registro, actualización de información de usuario, perfil, listar posts, agregar, actualizar y eliminar posts. Obtiene token de inicio de sesión y se valida temporalidad.",
        "contact": {
            "email": "contact@linuxitos.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/v1/post": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Obtener listado o un post específico",
                "operationId": "d7a7efb2146ca180ae57a075f0761e47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID del post a consultar. Si no se envía, devuelve la lista",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página (opcional)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad por página",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Texto de búsqueda",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Campo por el que ordenar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "asc o desc",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de registros"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/post/delete": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Eliminar un post",
                "operationId": "c155c0e641be82530e45e20e62901aef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "ID del post a eliminar",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post eliminado correctamente"
                    },
                    "404": {
                        "description": "Post no encontrado o no pertenece al usuario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/post/save": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Crear o actualizar un post",
                "operationId": "d4d42ea858c828b8f931938c80c1e392",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "ID del post. Si es 0 o no se envía, se crea uno nuevo.",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Mi primer post"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Contenido del post"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post creado o actualizado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}